LayoutTests/http/tests/referrer-policy-anchor/same-origin/cross-origin-http-http-expected.txt

Tests the behavior of same-origin referrer policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is expected
PASS successfullyParsed is true

TEST COMPLETE
